create useNavigateCommandMenu hook
