+++ packages/twenty-front/src/modules/command-menu/hooks/useCommandMenu.ts
@@ -1,9 +1,6 @@
 import { useRecoilCallback } from 'recoil';
 
 import { commandMenuSearchState } from '@/command-menu/states/commandMenuSearchState';
-import { useSelectableList } from '@/ui/layout/selectable-list/hooks/useSelectableList';
-import { usePreviousHotkeyScope } from '@/ui/utilities/hotkey/hooks/usePreviousHotkeyScope';
-import { AppHotkeyScope } from '@/ui/utilities/hotkey/types/AppHotkeyScope';
 import {
   IconBolt,
   IconCalendarEvent,
@@ -14,54 +11,31 @@ import {
 } from 'twenty-ui';
 
 import { COMMAND_MENU_COMPONENT_INSTANCE_ID } from '@/command-menu/constants/CommandMenuComponentInstanceId';
-import { COMMAND_MENU_CONTEXT_CHIP_GROUPS_DROPDOWN_ID } from '@/command-menu/constants/CommandMenuContextChipGroupsDropdownId';
 import { COMMAND_MENU_PREVIOUS_COMPONENT_INSTANCE_ID } from '@/command-menu/constants/CommandMenuPreviousComponentInstanceId';
 import { useCopyContextStoreStates } from '@/command-menu/hooks/useCopyContextStoreAndActionMenuStates';
-import { useResetContextStoreStates } from '@/command-menu/hooks/useResetContextStoreStates';
+import { useNavigateCommandMenu } from '@/command-menu/hooks/useNavigateCommandMenu';
 import { viewableRecordIdComponentState } from '@/command-menu/pages/record-page/states/viewableRecordIdComponentState';
 import { workflowIdComponentState } from '@/command-menu/pages/workflow/states/workflowIdComponentState';
-import { commandMenuNavigationMorphItemByPageState } from '@/command-menu/states/commandMenuNavigationMorphItemsState';
-import { commandMenuNavigationRecordsState } from '@/command-menu/states/commandMenuNavigationRecordsState';
-import { commandMenuNavigationStackState } from '@/command-menu/states/commandMenuNavigationStackState';
 import { commandMenuPageInfoState } from '@/command-menu/states/commandMenuPageInfoState';
-import { commandMenuPageState } from '@/command-menu/states/commandMenuPageState';
 import { hasUserSelectedCommandState } from '@/command-menu/states/hasUserSelectedCommandState';
 import { isCommandMenuClosingState } from '@/command-menu/states/isCommandMenuClosingState';
 import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
-import { MAIN_CONTEXT_STORE_INSTANCE_ID } from '@/context-store/constants/MainContextStoreInstanceId';
 import { contextStoreCurrentViewTypeComponentState } from '@/context-store/states/contextStoreCurrentViewTypeComponentState';
 import { contextStoreFiltersComponentState } from '@/context-store/states/contextStoreFiltersComponentState';
 import { contextStoreNumberOfSelectedRecordsComponentState } from '@/context-store/states/contextStoreNumberOfSelectedRecordsComponentState';
 import { contextStoreTargetedRecordsRuleComponentState } from '@/context-store/states/contextStoreTargetedRecordsRuleComponentState';
 import { ContextStoreViewType } from '@/context-store/types/ContextStoreViewType';
-import { viewableRecordIdState } from '@/object-record/record-right-drawer/states/viewableRecordIdState';
-import { useDropdownV2 } from '@/ui/layout/dropdown/hooks/useDropdownV2';
-import { emitRightDrawerCloseEvent } from '@/ui/layout/right-drawer/utils/emitRightDrawerCloseEvent';
 import { isDragSelectionStartEnabledState } from '@/ui/utilities/drag-select/states/internal/isDragSelectionStartEnabledState';
 import { t } from '@lingui/core/macro';
 import { useCallback } from 'react';
 import { v4 } from 'uuid';
 import { isCommandMenuOpenedState } from '../states/isCommandMenuOpenedState';
 
-export type CommandMenuNavigationStackItem = {
-  page: CommandMenuPages;
-  pageTitle: string;
-  pageIcon: IconComponent;
-  pageIconColor?: string;
-  pageId?: string;
-};
-
 export const useCommandMenu = () => {
-  const { resetSelectedItem } = useSelectableList('command-menu-list');
-  const {
-    setHotkeyScopeAndMemorizePreviousScope,
-    goBackToPreviousHotkeyScope,
-  } = usePreviousHotkeyScope();
+  const { navigateCommandMenu, onCommandMenuCloseAnimationComplete } =
+    useNavigateCommandMenu();
 
   const { copyContextStoreStates } = useCopyContextStoreStates();
-  const { resetContextStoreStates } = useResetContextStoreStates();
-
-  const { closeDropdown } = useDropdownV2();
 
   const closeCommandMenu = useRecoilCallback(
     ({ set }) =>
@@ -73,141 +47,7 @@ export const useCommandMenu = () => {
     [],
   );
 
-  const onCommandMenuCloseAnimationComplete = useRecoilCallback(
-    ({ set }) =>
-      () => {
-        closeDropdown(COMMAND_MENU_CONTEXT_CHIP_GROUPS_DROPDOWN_ID);
-
-        resetContextStoreStates(COMMAND_MENU_COMPONENT_INSTANCE_ID);
-        resetContextStoreStates(COMMAND_MENU_PREVIOUS_COMPONENT_INSTANCE_ID);
-
-        set(viewableRecordIdState, null);
-        set(commandMenuPageState, CommandMenuPages.Root);
-        set(commandMenuPageInfoState, {
-          title: undefined,
-          Icon: undefined,
-          instanceId: '',
-        });
-        set(isCommandMenuOpenedState, false);
-        set(commandMenuSearchState, '');
-        set(commandMenuNavigationMorphItemByPageState, new Map());
-        set(commandMenuNavigationRecordsState, []);
-        set(commandMenuNavigationStackState, []);
-        resetSelectedItem();
-        set(hasUserSelectedCommandState, false);
-        goBackToPreviousHotkeyScope();
-
-        emitRightDrawerCloseEvent();
-        set(isCommandMenuClosingState, false);
-      },
-    [
-      closeDropdown,
-      goBackToPreviousHotkeyScope,
-      resetContextStoreStates,
-      resetSelectedItem,
-    ],
-  );
-
-  const openCommandMenu = useRecoilCallback(
-    ({ snapshot, set }) =>
-      () => {
-        const isCommandMenuOpened = snapshot
-          .getLoadable(isCommandMenuOpenedState)
-          .getValue();
-
-        const isCommandMenuClosing = snapshot
-          .getLoadable(isCommandMenuClosingState)
-          .getValue();
-
-        if (isCommandMenuClosing) {
-          onCommandMenuCloseAnimationComplete();
-        }
-
-        setHotkeyScopeAndMemorizePreviousScope(AppHotkeyScope.CommandMenuOpen);
-
-        if (isCommandMenuOpened) {
-          return;
-        }
-
-        copyContextStoreStates({
-          instanceIdToCopyFrom: MAIN_CONTEXT_STORE_INSTANCE_ID,
-          instanceIdToCopyTo: COMMAND_MENU_COMPONENT_INSTANCE_ID,
-        });
-
-        set(isCommandMenuOpenedState, true);
-        set(hasUserSelectedCommandState, false);
-        set(isDragSelectionStartEnabledState, false);
-      },
-    [
-      copyContextStoreStates,
-      onCommandMenuCloseAnimationComplete,
-      setHotkeyScopeAndMemorizePreviousScope,
-    ],
-  );
-
-  const navigateCommandMenu = useRecoilCallback(
-    ({ snapshot, set }) => {
-      return ({
-        page,
-        pageTitle,
-        pageIcon,
-        pageIconColor,
-        pageId,
-        resetNavigationStack = false,
-      }: CommandMenuNavigationStackItem & {
-        resetNavigationStack?: boolean;
-      }) => {
-        if (!pageId) {
-          pageId = v4();
-        }
-
-        openCommandMenu();
-        set(commandMenuPageState, page);
-        set(commandMenuPageInfoState, {
-          title: pageTitle,
-          Icon: pageIcon,
-          instanceId: pageId,
-        });
-
-        const isCommandMenuClosing = snapshot
-          .getLoadable(isCommandMenuClosingState)
-          .getValue();
-
-        const currentNavigationStack = isCommandMenuClosing
-          ? []
-          : snapshot.getLoadable(commandMenuNavigationStackState).getValue();
-
-        if (resetNavigationStack) {
-          set(commandMenuNavigationStackState, [
-            {
-              page,
-              pageTitle,
-              pageIcon,
-              pageIconColor,
-              pageId,
-            },
-          ]);
-
-          set(commandMenuNavigationRecordsState, []);
-          set(commandMenuNavigationMorphItemByPageState, new Map());
-        } else {
-          set(commandMenuNavigationStackState, [
-            ...currentNavigationStack,
-            {
-              page,
-              pageTitle,
-              pageIcon,
-              pageIconColor,
-              pageId,
-            },
-          ]);
-        }
-      };
-    },
-    [openCommandMenu],
-  );
-
-  const openRootCommandMenu = useCallback(() => {
+  const openCommandMenu = useCallback(() => {
     navigateCommandMenu({
       page: CommandMenuPages.Root,
       pageTitle: 'Command Menu',
@@ -228,10 +68,10 @@ export const useCommandMenu = () => {
         if (isCommandMenuOpened) {
           closeCommandMenu();
         } else {
-          openRootCommandMenu();
+          openCommandMenu();
         }
       },
-    [closeCommandMenu, openRootCommandMenu],
+    [closeCommandMenu, openCommandMenu],
   );
 
   const openWorkflowTriggerTypeInCommandMenu = useRecoilCallback(
@@ -487,7 +327,7 @@ export const useCommandMenu = () => {
   );
 
   return {
-    openRootCommandMenu,
+    openCommandMenu,
     closeCommandMenu,
     onCommandMenuCloseAnimationComplete,
     navigateCommandMenu,
